requirements.txt
appdirs==1.4.3
awsebcli==3.10.0
backports.ssl-match-hostname==3.5.0.1
botocore==1.5.26
cement==2.8.2
colorama==0.3.7
Django==1.9.2
django-mongodb-engine==0.6.0
djangorestframework==3.3.0
djangotoolbox==1.8.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
jmespath==0.9.2
mongo==0.2.0
mongoengine==0.8.0
packaging==16.8
pathspec==0.5.0
pymongo==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.0
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
sql==0.4.0
sqlparse==0.2.3
tabulate==0.7.5
termcolor==1.1.0
websocket-client==0.40.0
MySQL-python==1.2.3